Refactor: 변수명 변경 및 폴더 이동
