+++ src/app/components/home/footer/index.tsx
