+++ src/app/components/home/footer/styles.ts
